requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
PyJWT==2.8.0
python-decouple==3.3
email-validator==2.2.0